backend/requirements.txt
flask
flask-cors
sqlalchemy
pandas
python-dateutil